Refactor V8 geometry filter
