+++ viewer/packages/cad-parsers/index.ts
@@ -25,7 +25,8 @@ export {
   filterPrimitivesOutsideClipBoxByCenterAndRadius,
   filterPrimitivesOutsideClipBoxByEllipse,
   filterPrimitivesOutsideClipBoxByVertices
-} from './src/cad/filterPrimitives';
+} from './src/cad/filterPrimitivesV8';
+
 
 export { filterInstanceMesh } from './src/cad/filterInstanceMesh';
